+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add background image for nicer loading and temp fix for minZoom issue
+- Add getRaster to TiffPyramidLoader
 
 ### Changed