+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add background image for nicer loading and temporary fix for minZoom issue
 - Add getRaster to TiffPyramidLoader
+- Picture-in-picture overview layer
 
 ### Changed